curso next seccion 13 completada

+++ src/todos/actions/todo-action.ts
@@ -3,6 +3,10 @@
 import prisma from "@/lib/prisma";
 import { Todo } from "@prisma/client";
 import { revalidatePath } from "next/cache";
+import { getServerSession } from "next-auth";
+import { authOptions } from "@/app/auth/authOptions";
+
+
 
 const sleep = (seconds: number) => {
     return new Promise(resolve => {
@@ -35,16 +39,26 @@ export const toggleTodo = async (id: string, complete: boolean): Promise<Todo> =
 export const createTodo = async (
   description: string,
   complete?: boolean
-): Promise<Todo> => {
-  const todo = await prisma.todo.create({
-    data: { description, complete },
-  });
-  revalidatePath("/dashboard/server-actions");
-  return todo;
+): Promise<Todo | null> => {
+   const session = await getServerSession(authOptions);
+  const userId = session?.user?.id;
+  let todo: Todo
+  if (userId) {
+      todo = await prisma.todo.create({
+      data: { description, complete, userId },
+    });
+    revalidatePath("/dashboard/server-actions");  
+      return todo;
+  }
+return null
+
 };
 
-export const deleteTodo = async ()=> {
-  const deleted = await prisma.todo.deleteMany({ where: { complete: true } });
+export const deleteTodo = async () => {
+  const session = await getServerSession(authOptions);
+  const userId = session?.user?.id;
+  const deleted = await prisma.todo.deleteMany({ where: { complete: true, userId } });
   revalidatePath("/dashboard/server-actions");
+  console.log(deleted)
   return deleted
 };
